+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * :fire: Hardwordx.java getDeviceModel rename to getModel and getDeviceName rename to getDevice
 * :fire: Hardwordx.kt removed getDeviceModel, getDeviceName, getHardware, getSupportedAbis method
 * :sparkles: Hardwordx's adds getProduct, getBrand method
+* :sparkles: Add SystemPropertiesx
 
 
 ## v0.4.0